# docs/swagger.yaml
basePath: /
definitions:
...
  dto.DeleteParty:
    properties:
      change_user:
        type: string
      party_code:
        type: integer
    type: object
...
  dto.PartyCode:
    properties:
      party_code:
        type: integer
    type: object
  dto.ResponseDto:
    properties:
      isSuccess:
        type: boolean
      message:
        type: string
      payload: {}
      statusCode:
        type: integer
    type: object
  dto.ResponseDtoV2:
    properties:
      data: {}
      isSuccess:
        type: boolean
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  dto.UpdateParty:
    properties:
      change_user:
        type: string
      id:
        type: integer
      party_address:
        type: string
      party_code:
        type: integer
      party_email:
        type: string
      party_name:
        type: string
      party_nid:
        type: string
      party_phone:
        type: string
    type: object
...
  model.ServiceParty:
    properties:
      id:
        type: integer
      party_address:
        type: string
      party_code:
        type: integer
      party_email:
        type: string
      party_name:
        type: string
      party_nid:
        type: string
      party_phone:
        type: string
    type: object
...
host: 118.67.213.45:7037
info:
  contact:
    email: suman@satcombd.com
    name: Sumon Sarker
  description: 'An API to perform `SICL IService API` operations. You can find out
    more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,
    #swagger](http://swagger.io/irc/).'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SICL IService API
  version: "2.0"
paths:
...
  /iservice/serive-party/get-service-party-id:
    get:
      description: Get New ServiceParty Id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseDtoV2'
      security:
      - BearerAuth: []
      summary: Get New ServiceParty Id
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/create-one-service-party:
    post:
      description: Create One ServiceParty
      parameters:
      - description: ServiceParty ID to get ServiceParty details
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/model.ServiceParty'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ResponseDtoV2'
            type: array
      security:
      - BearerAuth: []
      summary: Create One ServiceParty
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/delete-service-party:
    post:
      description: Create One ServiceParty
      parameters:
      - description: ServiceParty ID to get ServiceParty details
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteParty'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ResponseDto'
            type: array
      security:
      - BearerAuth: []
      summary: Create One ServiceParty
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/get-all-service-party:
    get:
      description: Get All ServiceParty
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseDtoV2'
      security:
      - BearerAuth: []
      summary: Get All ServiceParty
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/get-all-service-party-by-name:
    get:
      description: Get All ServiceParty
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseDtoV2'
      security:
      - BearerAuth: []
      summary: Get All ServiceParty
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/get-one-service-party:
    post:
      description: Get One ServiceParty
      parameters:
      - description: ServiceParty ID to get ServiceParty details
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/dto.PartyCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ResponseDtoV2'
            type: array
      security:
      - BearerAuth: []
      summary: Get One ServiceParty
      tags:
      - ServiceParty Entry Service
  /iservice/service-party/update-service-party:
    post:
      description: Update One ServiceParty
      parameters:
      - description: ServiceParty ID to get ServiceParty details
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateParty'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ResponseDtoV2'
            type: array
      security:
      - BearerAuth: []
      summary: Update One ServiceParty
      tags:
      - ServiceParty Entry Service
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"